refactor: add ChartLegend component
