+++ packages/ui/src/components/templates/Dashboard/Chart/index.tsx
@@ -18,4 +18,4 @@
 
 export {default as BarChart} from './BarChart';
 export {default as LineChart} from './LineChart';
-export {default as PieChart} from './PieChart';
+export {default as PieChart} from './PieChart/PieChart';
